Add a scripts module with a first example script
The goal of the scripts module is to provide a set of ready-to-use
functions for a set of concrete problems and situations. The first
script for example generates a csv file containing the PIG tables for
a given dataset, which can then be used to create PIG graphs for every
predictor in the dataset.