Change logic of CategoricalDataProcessor
Instead of keeping a list per column of categories to merge,
we now keep a list of categories to keep. This has the advantage
that if a new category arises in a test set (or a future dataset to score),
this will automatically be renamed to "Other" so that TargetEncoder can handle this.

Unit tests were also updated for this change