Add PreProcessor class as facade for preprocessing
The PreProcessor class unifies all the logic for preprocessing and
combines all preprocessing steps (which can be found in separate classes).

This means that the PreProcessor class is the main class to use from on.
Furthermore, it also has functionality to store the preprocessing pipeline
as a json file so that it can be easily reused and methods for train-test split.