Modify forward_selection algorithm logic
Change was made to make sure forced predictors
could be included in models with only positive coefficients
as much as possible. In the new implemtation, only when there
is no possible combination of forced predictors
(and other predictors) that lead to a model with only positive
coefficients will some of the forced predictors be skipped.