Bug fix in ForwardFeatureSelection.fit
The bug in the code was a subtle bug that did NOT affect
the correctness of the output (e.g. the best model).

In the fit function, the list of predictors was filtered
by removing the excluded predictors. However, the foreced predictors
where not filtered and this caused the following problem:
suppose you have 5 predcitors [a, b, c, d, e] and you want to
force three predictors [b, d, e] in the model. When you do the
forward selection, you have a couple of iterations
  - the first model with 1 predictor e.g. d
  - the seconds model with 2 predictors e.g. [b, d]
  - suppose now that in the 3rd iteration, the model with
    predictors [b, d, e] has no positive coeff (which we want to force
    for interpretability) then there is no model to be added here
  - in the next iteration, we have to go through the remaining list
    which in this case would be [a, c, e].

However, in the code implemented before, the list was [a, c, e, e]!
This means that here, a model with predictors [b, d, e] is trained twice!
Although this does not have an impact on the final result, an additional
model will be trained at every iteration which has an impact on
performance (computation). We expect a speed-up of cobra with this fix!